Make Build Possible
Needed to remove the modern option for Nuxt, because for some reason some modern files are then undefined and cause errors when loading. Additionally tsconfig paths were not resolved correctly, so the settings were changed to always use the path resolve plugin.
There still seems to be some kind of bug which happens sometimes when uploading images (which causes infinite loading), but I was not able to reproduce that and now only added a timeout after 60 seconds