Changed method of Icon font, see description
Due to the issue 341 in this repo https://github.com/nuxt-community/vuetify-module/ and this issue 311 https://github.com/nuxt-community/vuetify-module/ (no linking them directly to not make a cross reference on the other repo) (and me also experiencing the Maximum call stack size exceeded) I changed the method of the icons to a) no components, but class based and b) no more external vuetify config.

As tool https://icomoon.io/ was used. (Using the svg files which were already in the components)